ChatXiaomi/requirements.txt
rich
nodriver
python-dotenv
httpx
